Improve conformance with the Origami specification.
- Update status from active to maintained.
- Update code to pass `obt verify`.
- Update tests to run with `obt test`.
- Remove dev dependencies.
- Remove grunt task runner.
- Update the folder structure.
- Remove travis ci config.
- Add circle config. ftdomdelegate can now deploy to npm on tag.
