+++ .gitignore
@@ -1,5 +1,9 @@
-node_modules
-lcov-report
-lcov.info
-npm-debug.log
-/build
+.DS_Store
+.env
+/.sass-cache/
+/bower_components/
+/node_modules/
+/build/
+.idea/
+/demos/local
+/coverage
