--- .npmignore
@@ -1,5 +1,0 @@
-.github/
-test/
-.travis.yml
-bower.json
-GruntFile.js
