+++ bower.json
@@ -1,7 +1,7 @@
 {
   "name": "ftdomdelegate",
   "description": "Create and manage a DOM event delegator.",
-  "main": "lib/delegate.js",
+  "main": "main.js",
   "ignore": [".github", "test", ".npmignore", ".gitignore", "GruntFile.js"],
   "license": "MIT"
 }
