+++ package.json
@@ -1,40 +1,15 @@
 {
   "name": "ftdomdelegate",
-  "version": "3.0.0",
   "author": "FT Labs <enquiries@labs.ft.com> (http://labs.ft.com/)",
-  "description": "Create and manage a DOM event delegator.",
   "contributors": [
     "Matthew Caruana Galizia",
     "Sam Giles",
     "Matt Andrews"
   ],
-  "main": "lib/index.js",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/ftlabs/ftdomdelegate.git"
-  },
-  "engines": {
-    "node": "*"
-  },
-  "scripts": {
-    "test": "./node_modules/.bin/grunt && ./node_modules/.bin/buster-test"
-  },
   "keywords": [
     "delegate",
     "dom",
     "events"
   ],
-  "devDependencies": {
-    "buster": "~0.7.18",
-    "grunt": "~1.0.1",
-    "grunt-cli": "~1.2.0",
-    "grunt-buster": "~0.4.2",
-    "grunt-browserify": "5.0.0",
-    "grunt-contrib-uglify": "~2.0.0",
-    "istanbul": "*",
-    "buster-istanbul": "*",
-    "grunt-contrib-jshint": "~1.0.0"
-  },
-  "license": "MIT",
-  "homepage": "https://github.com/ftlabs/ftdomdelegate"
+  "license": "MIT"
 }
